Wikiscrape/requirements.txt
selenium
mysql-connector-python
pandas